:recycle: fix syntaxes for V2, bump test lib
