+++ mindee/parsing/v2/__init__.py
@@ -13,4 +13,4 @@ from mindee.parsing.v2.inference_options import InferenceOptions
 from mindee.parsing.v2.inference_response import InferenceResponse
 from mindee.parsing.v2.inference_result import InferenceResult
 from mindee.parsing.v2.job import Job
-from mindee.parsing.v2.polling_response import PollingResponse
+from mindee.parsing.v2.job_response import JobResponse
